Module scayl::context

The main module containing the scoring mechanism & deployment contexts

Structs

ContextRunner

The staging area for analyzing a piece of software

DeploymentContext

The context that describes the environment in which the vulnerability was found.

DeploymentScore

The score report of a single piece of software.

ScaylInfo

General information about a what scayl version generated a particular score.

VulnerabilityScore

The broken-down score of a single piece of software.

Enums

Error

An error that occurred while analyzing a piece of software

FileSystemAccess

Describes how much access the service has access to the file-system. This information is used in conjunction with information sensitivity to determine the effect of file-system exploits. Selection of these categories should fall on the highest value based on read + write. If a service cannot write to every file but can read every file, it should be given the “Full” identifier.

InformationSensitivity

The Information-Sensitivity category describes the sensitivity of the deployed service’s information. This category is used to analyze the effect of information-targeting vulnerabilities.

NetworkConfiguration

The Network Deployment category describes the public internet’s access to the deployed service. This category is used to evaluate the effect of a network-based vulnerability’s impact on the deployed service.

Permissions

Describes the level of commandline access this service has access to. This category is used to analyze the effect of code-execution vulnerabilities.

RemoteAccess

The Remote-Access category describes the remote access to the deployed service. This category is used to evaluate the effect of a remote-access vulnerability’s impact on the deployed service.

Traits

CvssProvider

A single cvss provider