Module scayl::format::sarif

Structs

Address

A physical or virtual address, or a range of addresses, in an ‘addressable region’ (memory or a binary file).

Artifact

A single artifact. In some cases, this artifact might be nested within another artifact.

ArtifactChange

A change to a single artifact.

ArtifactContent

The contents of the artifact.

ArtifactLocation

The location of the artifact.

Attachment

An artifact relevant to a result.

CodeFlow

A set of threadFlows which together describe a pattern of code execution relevant to detecting a result.

ConfigurationOverride

Information about how a specific rule or notification was reconfigured at runtime.

Conversion

A conversion object that will be merged with a separate run.

Edge

Represents a directed edge in a graph.

EdgeTraversal

Represents the traversal of a single edge during a graph traversal.

Exception

The runtime exception, if any, relevant to this notification.

ExternalProperties

The top-level element of an external property file.

ExternalPropertyFileReference

An external property file containing a run.conversion object to be merged with the root log file.

ExternalPropertyFileReferences

References to external property files that should be inlined with the content of a root log file.

Fix

A proposed fix for the problem represented by a result object. A fix specifies a set of artifacts to modify. For each artifact, it specifies a set of bytes to remove, and provides a set of new bytes to replace them.

Graph

A network of nodes and directed edges that describes some aspect of the structure of the code (for example, a call graph).

GraphTraversal

Represents a path through a graph.

Invocation

An invocation object that describes the invocation of the converter.

Location

The location to which this stack frame refers.

LocationRelationship

Information about the relation of one location to another.

LogicalLocation

A logical location of a construct that produced a result.

Message

A short description of the artifact.

MultiformatMessageString

An alternate rendered representation of the artifact (e.g., a decompiled representation of a binary region).

Node

Represents a node in a graph.

Notification

Describes a condition relevant to the tool itself, as opposed to being relevant to a target being analyzed by the tool.

PhysicalLocation

Identifies the artifact and region.

Rectangle

An area within an image.

Region

A region within an artifact where a result was detected.

Replacement

The replacement of a single region of an artifact.

ReportingConfiguration

Specifies how the rule or notification was configured during the scan.

ReportingDescriptor

Metadata that describes a specific report produced by the tool, as part of the analysis it provides or its runtime reporting.

ReportingDescriptorReference

A reference used to locate the descriptor whose configuration was overridden.

ReportingDescriptorRelationship

Information about the relation of one reporting descriptor to another.

ResultProvenance

Information about how and when the result was detected.

Run

Describes a single run of an analysis tool, and contains the reported output of that run.

RunAutomationDetails

Automation details that describe this run.

Sarif

Static Analysis Results Format (SARIF) Version 2.1.0 JSON Schema: a standard format for the output of static analysis tools.

SpecialLocations

A specialLocations object that defines locations of special significance to SARIF consumers.

Stack

The sequence of function calls leading to the exception.

StackFrame

A function call within a stack trace.

Suppression

A suppression that is relevant to a result.

ThreadFlow

Describes a sequence of code locations that specify a path through a single thread of execution such as an operating system or fiber.

ThreadFlowLocation

A location visited by an analysis tool while simulating or monitoring the execution of a program.

Tool

A tool object that describes the converter.

ToolComponent

The analysis tool that was run.

ToolComponentReference

A reference used to locate the toolComponent associated with the descriptor.

TranslationMetadata

Translation metadata, required for a translation, not populated by other component types.

TrivyResult

A result produced by an analysis tool.

VersionControlDetails

Specifies the information necessary to retrieve a desired revision from a version control system.

WebRequest

A web request associated with this thread flow location.

WebResponse

A web response associated with this thread flow location.

Enums

BaselineState

The state of a result relative to a baseline of a previous run.

ColumnKind

Specifies the unit in which the tool measures columns.

Content

Importance

Specifies the importance of this location in understanding the code flow in which it occurs. The order from most to least important is “essential”, “important”, “unimportant”. Default: “important”.

Level

Specifies the failure level for the report.

ResultKind

A value that categorizes results by evaluation state.

Role

State

A string that indicates the state of the suppression.

SuppressionKind

A string that indicates where the suppression is persisted.

Version

The SARIF format version of this external properties object.

Type Definitions

PropertyBag

Key/value pairs that provide additional information about the address.