Type Definition scayl::format::sarif::PropertyBag

pub type PropertyBag = BTreeMap<String, Value>;

Key/value pairs that provide additional information about the address.

Key/value pairs that provide additional information about the object.

Key/value pairs that provide additional information about the artifact content.

Key/value pairs that provide additional information about the message.

Key/value pairs that provide additional information about the artifact location.

Key/value pairs that provide additional information about the artifact.

Contains configuration information specific to a report.

Key/value pairs that provide additional information about the reporting configuration.

Key/value pairs that provide additional information about the reporting descriptor reference.

Key/value pairs that provide additional information about the toolComponentReference.

Key/value pairs that provide additional information about the configuration override.

Key/value pairs that provide additional information about the invocation.

Key/value pairs that provide additional information about the exception.

Key/value pairs that provide additional information about the region.

Key/value pairs that provide additional information about the logical location.

Key/value pairs that provide additional information about the physical location.

Key/value pairs that provide additional information about the location.

Key/value pairs that provide additional information about the location relationship.

Key/value pairs that provide additional information about the stack frame.

Key/value pairs that provide additional information about the stack.

Key/value pairs that provide additional information about the notification.

Key/value pairs that provide additional information about the conversion.

Key/value pairs that provide additional information about the report.

Key/value pairs that provide additional information about the tool component.

Key/value pairs that provide additional information about the translation metadata.

Key/value pairs that provide additional information about the tool.

Key/value pairs that provide additional information that will be merged with a separate run.

Key/value pairs that provide additional information about the edge.

Key/value pairs that provide additional information about the node.

Key/value pairs that provide additional information about the graph.

Key/value pairs that provide additional information about the external properties.

Key/value pairs that provide additional information about the attachment.

Key/value pairs that provide additional information about the rectangle.

Key/value pairs that provide additional information about the code flow.

Key/value pairs that provide additional information about the threadflow location.

Key/value pairs that provide additional information about the request.

Key/value pairs that provide additional information about the response.

Key/value pairs that provide additional information about the thread flow.

Key/value pairs that provide additional information about the change.

Key/value pairs that provide additional information about the replacement.

Key/value pairs that provide additional information about the fix.

Key/value pairs that provide additional information about the edge traversal.

Key/value pairs that provide additional information about the graph traversal.

Key/value pairs that provide additional information about the result.

Key/value pairs that provide additional information about the suppression.

Key/value pairs that provide additional information about the log file.

Key/value pairs that provide additional information about the run automation details.

Key/value pairs that provide additional information about the external property file.

Key/value pairs that provide additional information about the external property files.

Key/value pairs that provide additional information about the run.

Key/value pairs that provide additional information about the special locations.

Key/value pairs that provide additional information about the version control details.